requirements.txt
rocketry~=2.5.1
pydantic~=1.10.0
httpx~=0.25.0
cryptography~=40.0.2
tortoise-orm[asyncpg]
python-dotenv~=1.0.0
asyncpg
httpx[socks]